change: state of health is uint8, nog uint16.

+++ App/BatMon_BQ35100/BatMon_functions.h
@@ -14,7 +14,7 @@ typedef struct {
     int16_t voltage;
     int16_t current;
     int16_t temperature;
-    uint16_t stateOfHealth;
+    uint8_t stateOfHealth;
     uint16_t ScaledR;
     uint16_t MeasuredZ;
